+++ src/common/dto/series-item.dto.ts
@@ -1,0 +1,10 @@
+export class SeriesItemDto {
+  name: string;
+
+  data: number[];
+
+  constructor(name: string, data: number[]) {
+    this.name = name;
+    this.data = data;
+  }
+}