+++ src/common/dto/submission-by-category.dto.ts
@@ -1,0 +1,5 @@
+export class SubmissionByCategoryDto {
+  state: string;
+
+  count: number;
+}