+++ src/modules/document/document.service.ts
@@ -1,5 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { StorageService } from '@nhogs/nestjs-firebase';
+import moment from 'moment';
 import { Repository } from 'typeorm';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
 
@@ -10,12 +12,21 @@ import { FileNotPdfException } from '../../exceptions';
 import type { IFile } from '../../interfaces';
 import { AwsS3Service } from '../../shared/services/aws-s3.service';
 import { ValidatorService } from '../../shared/services/validator.service';
+import { StatisticsDto } from '../staff/dto/statistics.dto';
 import { StaffService } from '../staff/staff.service';
 import { WorkflowService } from '../workflow/workflow.service';
 import { CreateDocumentDto } from './dto/create-document.dto';
 import type { DocumentDto } from './dto/document.dto';
 import type { UpdateDocumentDto } from './dto/update-document.dto';
 import { DocumentEntity } from './entities/document.entity';
+
+const uniqueStatuses = [
+  DocumentState.PENDING,
+  DocumentState.APPROVED,
+  DocumentState.REJECTED,
+  DocumentState.CHANGE_REQUESTED,
+  DocumentState.DRAFT,
+];
 
 @Injectable()
 export class DocumentService {
@@ -26,6 +37,7 @@ export class DocumentService {
     private workflowService: WorkflowService,
     private validatorService: ValidatorService,
     private awsS3Service: AwsS3Service,
+    private storageService: StorageService,
   ) {}
 
   @Transactional()
@@ -265,7 +277,12 @@ export class DocumentService {
     }
 
     if (file) {
-      docEntity.reviewerAttachment = await this.awsS3Service.uploadImage(file);
+      const fileName = `${docEntity.title}_${userId}_${file.originalname}`;
+      await this.storageService.uploadBytes(fileName, file.buffer);
+
+      docEntity.reviewerAttachment = await this.storageService.getDownloadURL(
+        fileName,
+      );
     }
 
     const workflow = await this.workflowService.findOne(docEntity.workflowId);
@@ -274,7 +291,7 @@ export class DocumentService {
       (x) => x.groupRole.designation === designation,
     );
 
-    if (currentIdx < --workflow.workflowItems.length) {
+    if (currentIdx < workflow.workflowItems.length - 1) {
       const staffEntity = await ([
         StaffDesignation.HOD,
         StaffDesignation.Dean,
@@ -291,6 +308,8 @@ export class DocumentService {
       if (!staffEntity) {
         throw new NotFoundException('Staff not found');
       }
+
+      docEntity.currentlyAssignedId = staffEntity.id;
 
       await this.docRepository.update(
         { id: docId },
@@ -511,7 +530,7 @@ export class DocumentService {
    * It returns a document entity with the specified id, owned by the specified student, with the owner, currently assigned
    * and workflow details
    * @param {Uuid} studentId - Uuid - This is the id of the student who owns the document.
-   * @param {Uuid} id - Uuid - This is the id of the document you want to retrieve.
+   * @param {Uuid} docId - Uuid - This is the id of the document you want to retrieve.
    * @returns A document entity
    */
   async studentFindOne(studentId: Uuid, docId: Uuid): Promise<DocumentEntity> {
@@ -621,5 +640,139 @@ export class DocumentService {
     );
 
     return docEntity;
+  }
+
+  /**
+   * It returns the number of documents in the database
+   * @returns The number of documents in the database.
+   */
+  async getCount(): Promise<number> {
+    const queryBuilder = this.docRepository.createQueryBuilder('doc');
+
+    return queryBuilder.getCount();
+  }
+
+  /**
+   * It returns the total number of documents, the number of students who have documents, and the number of documents in
+   * each state
+   * @param {Uuid} staffId - Uuid - The staff's id
+   * @returns An object with the total number of documents, the number of students, and the number of documents in each
+   * category.
+   */
+  async getStaffStatistics(staffId: Uuid): Promise<StatisticsDto> {
+    const startOfDay = new Date();
+    startOfDay.setDate(startOfDay.getDate() - 7);
+    startOfDay.setHours(0, 0, 0, 0);
+
+    const queryBuilder = this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.currentlyAssigned = :id', { id: staffId });
+
+    const total = await queryBuilder.getCount();
+    const studentCount = await queryBuilder
+      .select('COUNT(DISTINCT(owner_id))', 'count')
+      .getRawOne();
+    const categories = await this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.currentlyAssigned = :id', { id: staffId })
+      .groupBy('doc.state')
+      .orderBy('doc.state')
+      .select('COUNT(doc.id)', 'count')
+      .addSelect('doc.state', 'state')
+      .getRawMany();
+
+    const start = moment().subtract(6, 'days').startOf('day');
+    const end = moment().endOf('day');
+
+    const queryResult = await this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.currentlyAssigned = :id', { id: staffId })
+      .select("state, DATE_TRUNC('day', created_at) as day, COUNT(id) as count")
+      .where('created_at BETWEEN :start AND :end', { start, end })
+      .groupBy('state, day')
+      .orderBy('day, state')
+      .getRawMany();
+
+    const processedData = uniqueStatuses.map((state) => {
+      const itemsByStatus = queryResult.filter((row) => row.state === state);
+      // eslint-disable-next-line unicorn/no-array-reduce
+      const countByDay = itemsByStatus.reduce((acc, row) => {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+        const day = moment(row.day).diff(start, 'days');
+        acc[day] = Number(row.count);
+
+        return acc;
+      }, Array.from({ length: 7 }).fill(0));
+
+      return { name: state, data: countByDay };
+    });
+
+    const stats = new StatisticsDto();
+
+    stats.submissions = total;
+    stats.submissionsByCategory = categories;
+    stats.students = Number(studentCount.count);
+    stats.weeklySubmissions = processedData;
+
+    return stats;
+  }
+
+  /**
+   * It returns a statistics object for a given student
+   * @param {Uuid} studentId - Uuid - The student's ID
+   * @returns A StatisticsDto object
+   */
+  async getStudentStatistics(studentId: Uuid): Promise<StatisticsDto> {
+    const queryBuilder = this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.ownerId = :id', { id: studentId });
+
+    const total = await queryBuilder.getCount();
+    const staffCount = await queryBuilder
+      .select('COUNT(DISTINCT(currently_assigned_id))', 'count')
+      .getRawOne();
+    const categories = await this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.ownerId = :id', { id: studentId })
+      .groupBy('doc.state')
+      .orderBy('doc.state')
+      .select('COUNT(doc.id)', 'count')
+      .addSelect('doc.state', 'state')
+      .getRawMany();
+
+    const start = moment().subtract(6, 'days').startOf('day');
+    const end = moment().endOf('day');
+
+    const queryResult = await this.docRepository
+      .createQueryBuilder('doc')
+      .where('doc.ownderId = :id', { id: studentId })
+      .select("state, DATE_TRUNC('day', created_at) as day, COUNT(id) as count")
+      .where('created_at BETWEEN :start AND :end', { start, end })
+      .groupBy('state, day')
+      .orderBy('day, state')
+      .getRawMany();
+
+    const processedData = uniqueStatuses.map((state) => {
+      const itemsByStatus = queryResult.filter((row) => row.state === state);
+      // eslint-disable-next-line unicorn/no-array-reduce
+      const countByDay = itemsByStatus.reduce((acc, row) => {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+        const day = moment(row.day).diff(start, 'days');
+        acc[day] = Number(row.count);
+
+        return acc;
+      }, Array.from({ length: 7 }).fill(0));
+
+      return { name: state, data: countByDay };
+    });
+
+    const stats = new StatisticsDto();
+
+    stats.submissions = total;
+    stats.submissionsByCategory = categories;
+    stats.students = Number(staffCount.count);
+    stats.weeklySubmissions = processedData;
+
+    return stats;
   }
 }