+++ src/shared/shared.module.ts
@@ -6,6 +6,7 @@ import { CqrsModule } from '@nestjs/cqrs';
 import { ApiConfigService } from './services/api-config.service';
 import { AwsS3Service } from './services/aws-s3.service';
 import { GeneratorService } from './services/generator.service';
+import { NotificationService } from './services/notification.service';
 import { TranslationService } from './services/translation.service';
 import { ValidatorService } from './services/validator.service';
 
@@ -15,6 +16,7 @@ const providers = [
   AwsS3Service,
   GeneratorService,
   TranslationService,
+  NotificationService,
   // {
   //   provide: 'NATS_SERVICE',
   //   useFactory: (configService: ApiConfigService) => {